Use importorskip instead where appropriate.
This makes it explicit why certain tests are optional. Also they will
automatically be run if the dependencies are available without the
requirement of having the dependencies of all optional tests installed.
The `-rs` option to pytest allows to display the reason for skipped
tests which gives an easy diagnosis which tests are missing.